Move firebase initialization into index
